requirements.txt
flask
opencv-python
numpy
werkzeug
flask
opencv-python
pytesseract
numpy
flask
pytesseract
Pillow
opencv-python
numpy
werkzeug
